webnn/relu-tests_gpu.https.html

<!doctype html>
<meta charset=utf-8>
<title>sample test WebNN API relu operation on GPU device</title>
<meta name="timeout" content="long">
<script>
self.GLOBAL = {
  isWindow: function() { return true; },
  isWorker: function() { return false; },
  isShadowRealm: function() { return false; },
};
</script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/utils.js"></script>
<div id=log></div>
<script>
  let context;
  let builder;
  promise_setup(async () => {
    context = await navigator.ml.createContext({deviceType: 'gpu'});
    builder = new MLGraphBuilder(context);
  }, {explicit_timeout: true, test_timeout: 180000});
  promise_test(async () => {
    const input = builder.input('x', {type: 'float32', dimensions: [24]});
    const output = builder.relu(input);
    // asynchronously compile the graph up to the output operand
    const graph = await builder.build({output});
    const inputs = {
      'x': new Float32Array([
        79.04725231657116,
        2.2503608756501166,
        80.73939090529203,
        63.90392076789547,
        77.67340745512104,
        -71.09157819044825,
        -82.74703468427575,
        -26.814426600801028,
        -99.16788836397058,
        -35.71083406288831,
        18.36165830990626,
        -37.36091648205435,
        -52.8386119809162,
        -10.408374773008958,
        60.60291560236189,
        -13.644198913810342,
        -76.54250291031946,
        -8.132338049258351,
        51.51447452437017,
        -51.63370281825297,
        -64.567999424324,
        -5.093302411117136,
        15.354103550744384,
        90.03858807393516
      ])
    };
    const outputs = {'output': new Float32Array(24)};
    // asynchronously execute the compiled graph
    const result = await context.compute(graph, inputs, outputs);
    const expected = [
      79.04724884033203,
      2.2503609657287598,
      80.73938751220703,
      63.9039192199707,
      77.67340850830078,
      0,
      0,
      0,
      0,
      0,
      18.361658096313477,
      0,
      0,
      0,
      60.6029167175293,
      0,
      0,
      0,
      51.51447296142578,
      0,
      0,
      0,
      15.354103088378906,
      90.03858947753906
    ];
    assert_array_approx_equals_ulp(result.outputs.output, expected, 0, 'relu', 'test WebNN relu for float32');
  }, 'test WebNN API relu operation on GPU device');
</script>
